srv/mycrm/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
django-allauth==65.10.0
fritzconnection==1.15.0
idna==3.10
msal==1.33.0
mysqlclient==2.2.7
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
PyYAML>=6,<7
requests-oauthlib>=1.3.1,<3
oauthlib>=3.2.2
